rename reaction force sensor name
