+++ tests/finite_element_problem/test_am_layers.py
@@ -6,7 +6,7 @@ import pytest
 from fenicsxconcrete.experimental_setup.am_multiple_layers import AmMultipleLayers
 from fenicsxconcrete.finite_element_problem.linear_elasticity import LinearElasticity
 from fenicsxconcrete.helper import Parameters
-from fenicsxconcrete.sensor_definition.other_sensor import ReactionForceSensorBottom
+from fenicsxconcrete.sensor_definition.reaction_force_sensor import ReactionForceSensorBottom
 from fenicsxconcrete.unit_registry import ureg
